requirements.txt
selenium==4.20.0
python-telegram-bot==20.3
